moved trainer into its own module
